Upgrade to Bio-Formats 5.7.1
This upgrades the parent POM to pom-scijava 17.1.1.
It enables the Maven enforcer, to ensure dependencies are sane.

It updates uses of the Bio-Formats API from 4.x to 5.x. However, I do
not know how to test, so am not sure whether this work is 100% correct.

The reason I did this work is because I was trying to enable the
enforcer, and it pointed out incompatibilities between the use of
the old Bio-Formats 4.x API compared to the SCIFIO artifacts,
which require Bio-Formats 5.x. However, it turns out that SCIFIO
is not actually used by this project.